查看 Instance Reference 的更多版本

+++ src/Polymerium.Trident/Models/CurseForgeApi/Pagination.cs
@@ -1,3 +1,3 @@
 ﻿namespace Polymerium.Trident.Models.CurseForgeApi;
 
-public record Pagination(uint Index, uint PageSize, uint ResultCount, ulong TotalCount);
+public record Pagination(uint Index, uint PageSize, uint ResultCount, uint TotalCount);
